Entrega 3 terminada: Inmueble(Listar, actualizar estado y cargar nuevo)

+++ app/src/main/java/com/miguel/inmobile/modelo/Inmueble.java
@@ -8,7 +8,7 @@ public class Inmueble implements Serializable {
     private String tipo;
     private String uso;
     private int ambientes;
-    private double superficie;
+    private int superficie;
     private double latitud;
 
     private double valor;
@@ -21,7 +21,7 @@ public class Inmueble implements Serializable {
     public Inmueble() {
     }
 
-    public Inmueble(int idInmueble, String direccion, String tipo, String uso, int ambientes, double superficie, double latitud, double valor, boolean disponible, String imagen, double longitud, int idPropietario, Propietario duenio) {
+    public Inmueble(int idInmueble, String direccion, String tipo, String uso, int ambientes, int superficie, double latitud, double valor, boolean disponible, String imagen, double longitud, int idPropietario, Propietario duenio) {
         this.idInmueble = idInmueble;
         this.direccion = direccion;
         this.tipo = tipo;
@@ -77,11 +77,11 @@ public class Inmueble implements Serializable {
         this.ambientes = ambientes;
     }
 
-    public double getSuperficie() {
+    public int getSuperficie() {
         return superficie;
     }
 
-    public void setSuperficie(double superficie) {
+    public void setSuperficie(int superficie) {
         this.superficie = superficie;
     }
